+++ noodles-cram/src/crai/writer.rs
@@ -68,13 +68,14 @@ where
     ///
     /// ```
     /// use noodles_bam as bam;
+    /// use noodles_core::Position;
     /// use noodles_cram::crai;
     ///
     /// let mut writer = crai::Writer::new(Vec::new());
     ///
     /// let index = vec![crai::Record::new(
     ///     Some(bam::record::ReferenceSequenceId::from(0)),
-    ///     10946,
+    ///     Position::new(10946),
     ///     6765,
     ///     17711,
     ///     233,
@@ -114,10 +115,11 @@ mod tests {
     #[test]
     fn test_write_record() -> Result<(), Box<dyn std::error::Error>> {
         use noodles_bam as bam;
+        use noodles_core::Position;
 
         let index = vec![Record::new(
             Some(bam::record::ReferenceSequenceId::from(0)),
-            10946,
+            Position::new(10946),
             6765,
             17711,
             233,